+++ tests/Unit/Manager/SimpleJobManagerTest.php
@@ -14,7 +14,6 @@ final class SimpleJobManagerTest extends TestCase
 	public function test(): void
 	{
 		$manager = new SimpleJobManager();
-		self::assertSame([], $manager->getExpressions());
 		self::assertSame([], $manager->getJobSchedules());
 		self::assertNull($manager->getJobSchedule(0));
 		self::assertNull($manager->getJobSchedule('id'));
@@ -30,13 +29,6 @@ final class SimpleJobManagerTest extends TestCase
 		$expression2 = clone $expression1;
 		$manager->addJob($job2, $expression2, 'id', 1);
 
-		self::assertSame(
-			[
-				0 => $expression1,
-				'id' => $expression2,
-			],
-			$manager->getExpressions(),
-		);
 		self::assertEquals(
 			[
 				0 => JobSchedule::create($job1, $expression1, 0),