+++ CHANGELOG.md
@@ -43,6 +43,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	- acquired job locks are scoped just to their id - changing run frequency or job name will not make process loose
 	  the lock
 
+### Removed
+
+- `JobManager`
+	- `getExpressions()` - replaced by `getJobSchedules()`
+
 ### Fixed
 
 - `ListCommand`